refactor: UTC 시간대 통일 + MemberTypingStats 개선
- JVM 타임존 UTC 설정 (TypingPracticeBeApplication)
- BaseEntity, MemberTypingStats, QuoteSimilarityRejectLog의
  LocalDateTime 컬럼에 timestamp with time zone 적용
- MemberTypingStats.totalPracticeTimeMin long → float 변경
- MemberTypingAggregation.totalPracticeTimeMin long → float 변경
- refreshForMember 중복 merge 방지
  - updatedAt 기반으로 마지막 갱신 이후 기록만 집계
- countAllByQuoteIds Number 타입 캐스팅 방어

+++ typing-practice-be/src/main/java/com/typingpractice/typing_practice_be/typingrecord/statistics/dto/MemberTypingAggregation.java
@@ -13,7 +13,7 @@ public class MemberTypingAggregation {
   private double avgCpm;
   private double avgAcc;
   private int bestCpm;
-  private long totalPracticeTimeMin;
+  private float totalPracticeTimeMin;
   private int totalResetCount;
   private LocalDateTime lastPracticedAt;
 }
